+++ Sources/MapboxMaps/MapView/MapView+Supportable.swift
@@ -32,8 +32,8 @@ extension MapView: OrnamentSupportableView {
 
 extension MapView: LocationSupportableMapView {
 
-    public func screenCoordinate(for locationCoordinate: CLLocationCoordinate2D) -> ScreenCoordinate {
-        return mapboxMap.__map.pixelForCoordinate(for: locationCoordinate)
+    public func point(for coordinate: CLLocationCoordinate2D) -> CGPoint {
+        return mapboxMap.point(for: coordinate)
     }
 
     public func metersPerPointAtLatitude(latitude: CLLocationDegrees) -> CLLocationDistance {