+++ Tests/MapboxMapsTests/Location/LocationSupportableMapViewMock.swift
@@ -14,16 +14,16 @@ import UIKit
 
 class LocationSupportableMapViewMock: UIView, LocationSupportableMapView {
 
+    public func point(for coordinate: CLLocationCoordinate2D) -> CGPoint {
+        return .zero
+    }
+
     func subscribeRenderFrameHandler(_ handler: @escaping (MapboxCoreMaps.Event) -> Void) {
         print("Pass through implementation")
     }
 
     func subscribeStyleChangeHandler(_ handler: @escaping (MapboxCoreMaps.Event) -> Void) {
         print("Pass through implementation")
-    }
-
-    func screenCoordinate(for locationCoordinate: CLLocationCoordinate2D) -> ScreenCoordinate {
-        return ScreenCoordinate(x: 0.0, y: 0.0)
     }
 
     func metersPerPointAtLatitude(latitude: CLLocationDegrees) -> CLLocationDistance {