+++ Tests/MapboxMapsTests/MapView/Integration Tests/Map/ObservableIntegrationTests.swift
@@ -45,7 +45,7 @@ class ObservableIntegrationTests: MapViewIntegrationTestCase {
             eventExpectation.fulfill()
         }
 
-        mapView.mapboxMap.__map.subscribe(for: observer, events: ["resource-request"])
+        mapView.mapboxMap.subscribe(observer, events: ["resource-request"])
 
         style.uri = .streets
 
@@ -57,6 +57,6 @@ class ObservableIntegrationTests: MapViewIntegrationTestCase {
 
         wait(for: [styleLoadExpectation, eventExpectation], timeout: 5.0)
 
-        mapView.mapboxMap.__map.unsubscribe(for: observer)
+        mapView.mapboxMap.unsubscribe(observer)
     }
 }